Renames `dump` to `cat`, preserves `dump` alias
