+++ src/bin/ion/commands/beta/from/json.rs
@@ -1,6 +1,8 @@
-use crate::commands::{dump, IonCliCommand, WithIonCliArgument};
 use anyhow::Result;
 use clap::{ArgMatches, Command};
+
+use crate::commands::cat::CatCommand;
+use crate::commands::{IonCliCommand, WithIonCliArgument};
 
 pub struct FromJsonCommand;
 
@@ -17,9 +19,9 @@ impl IonCliCommand for FromJsonCommand {
         command.with_input().with_output().with_format()
     }
 
-    fn run(&self, _command_path: &mut Vec<String>, args: &ArgMatches) -> Result<()> {
-        // Because JSON data is valid Ion, the `dump` command may be reused for converting JSON.
+    fn run(&self, command_path: &mut Vec<String>, args: &ArgMatches) -> Result<()> {
+        // Because JSON data is valid Ion, the `cat` command may be reused for converting JSON.
         // TODO ideally, this would perform some smarter "up-conversion".
-        dump::run("json", args)
+        CatCommand.run(command_path, args)
     }
 }
